[Cosmetic] Move specific ShapeShapeDistance implementation into subdir.
