+++ src/distance/sphere_plane.cpp
@@ -40,8 +40,9 @@
 #include <limits>
 #include <hpp/fcl/math/transform.h>
 #include <hpp/fcl/shape/geometric_shapes.h>
-#include <../src/distance_func_matrix.h>
-#include "narrowphase/details.h"
+
+#include "../distance_func_matrix.h"
+#include "../narrowphase/details.h"
 
 namespace hpp
 {
@@ -49,15 +50,15 @@ namespace fcl {
   struct GJKSolver;
 
   template <>
-  FCL_REAL ShapeShapeDistance <Cone, Halfspace>
+  FCL_REAL ShapeShapeDistance <Sphere, Plane>
   (const CollisionGeometry* o1, const Transform3f& tf1,
    const CollisionGeometry* o2, const Transform3f& tf2,
    const GJKSolver*, const DistanceRequest&,
    DistanceResult& result)
   {
-    const Cone& s1 = static_cast <const Cone&> (*o1);
-    const Halfspace& s2 = static_cast <const Halfspace&> (*o2);
-    details::coneHalfspaceIntersect
+    const Sphere& s1 = static_cast <const Sphere&> (*o1);
+    const Plane& s2 = static_cast <const Plane&> (*o2);
+    details::spherePlaneIntersect
       (s1, tf1, s2, tf2, result.min_distance, result.nearest_points [0],
        result.nearest_points [1], result.normal);
     result.o1 = o1; result.o2 = o2; result.b1 = -1; result.b2 = -1;
@@ -65,15 +66,15 @@ namespace fcl {
   }
 
   template <>
-  FCL_REAL ShapeShapeDistance <Halfspace, Cone>
+  FCL_REAL ShapeShapeDistance <Plane, Sphere>
   (const CollisionGeometry* o1, const Transform3f& tf1,
    const CollisionGeometry* o2, const Transform3f& tf2,
    const GJKSolver*, const DistanceRequest&,
    DistanceResult& result)
   {
-    const Halfspace& s1 = static_cast <const Halfspace&> (*o1);
-    const Cone& s2 = static_cast <const Cone&> (*o2);
-    details::coneHalfspaceIntersect
+    const Plane& s1 = static_cast <const Plane&> (*o1);
+    const Sphere& s2 = static_cast <const Sphere&> (*o2);
+    details::spherePlaneIntersect
       (s2, tf2, s1, tf1, result.min_distance, result.nearest_points [1],
        result.nearest_points [0], result.normal);
     result.o1 = o1; result.o2 = o2; result.b1 = -1; result.b2 = -1;
